requirements.txt
cfgv==3.4.0
coverage==7.3.4
distlib==0.3.8
filelock==3.13.1
identify==2.5.33
nodeenv==1.8.0
platformdirs==4.1.0
pre-commit==3.6.0
python-dotenv==1.0.0
PyYAML==6.0.1
ruff==0.1.8
setuptools==69.0.2
virtualenv==20.25.0
